rs, sys: bump capstone to 6.0.0-Alpha6

+++ capstone-rs/src/arch/alpha.rs
@@ -48,7 +48,7 @@ pub enum AlphaOperandType {
     Reg(RegId),
 
     /// Immediate
-    Imm(i32),
+    Imm(i64),
 
     /// Invalid
     Invalid,
